update: 更新文件 tcp_server_send_context.cpp

+++ frameworks/js/napi/socket/async_context/src/tcp_server_send_context.cpp
@@ -67,7 +67,7 @@ int TcpServerSendContext::GetSocketFd() const
     if (sharedManager_ == nullptr) {
         return -1;
     }
-    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<intptr_t>(sharedManager_->GetData())) : -1;
+    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<uint64_t>(sharedManager_->GetData())) : -1;
 }
 
 bool TcpServerSendContext::CheckParamsType(napi_value *params, size_t paramsCount)
